requirements.txt
torch>=2.0.0
numpy
pandas
scikit-learn
pillow
albumentations>=1.3.0
timm
transformers>=4.30.0
matplotlib
ipykernel